+++ app/src/main/java/com/ss20/se2/monopoly/models/fields/cards/ChanceCard.java
@@ -2,7 +2,13 @@ package com.ss20.se2.monopoly.models.fields.cards;
 
 public class ChanceCard extends Card{
 
-	public ChanceCard(String name){
+	private String description;
+
+	public ChanceCard(String name, String description){
 		super(name);
+		this.description = description;
+	}
+	public String getDescription(){
+		return description;
 	}
 }
