+++ app/src/main/java/com/ss20/se2/monopoly/models/fields/cards/CommunityCard.java
@@ -2,7 +2,14 @@ package com.ss20.se2.monopoly.models.fields.cards;
 
 public class CommunityCard extends Card{
 
-	public CommunityCard(String name){
+	private String description;
+
+
+	public CommunityCard(String name, String description){
 		super(name);
+		this.description = description;
+	}
+	public String getDescription(){
+		return description;
 	}
 }
