--- app/src/main/java/com/ss20/se2/monopoly/view/OldActivity2.java
@@ -1,206 +1,0 @@
-package com.ss20.se2.monopoly.view;
-
-import android.widget.ImageView;
-import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.DialogInterface;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.TextView;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.ss20.se2.monopoly.R;
-import com.ss20.se2.monopoly.models.*;
-import com.ss20.se2.monopoly.models.fields.GameTile;
-import com.ss20.se2.monopoly.models.fields.cards.ChanceCard;
-import com.ss20.se2.monopoly.models.fields.cards.CommunityCard;
-import com.ss20.se2.monopoly.models.fields.deeds.Deed;
-import com.ss20.se2.monopoly.models.fields.deeds.Railroad;
-import com.ss20.se2.monopoly.models.fields.deeds.Street;
-import com.ss20.se2.monopoly.models.fields.deeds.Utility;
-
-public class OldActivity2 extends AppCompatActivity{
-
-	Button button_rollDice;
-	TextView view_numberDice;
-	TextView view_position;
-	TextView view_balance;
-
-	Dice dice = new Dice();
-	Gameboard gameboard;
-	ChanceCardDeck chanceCards;
-	CommunityCardDeck communityCards;
-	ChanceCardProcessor chanceCardProcessor;
-	int amount;
-
-	public ImageView[] initializeUI(){
-		ImageView field0 = findViewById(R.id.tile_0);
-		ImageView field1 = findViewById(R.id.tile_1);
-		ImageView field2 = findViewById(R.id.tile_2);
-		ImageView field3 = findViewById(R.id.tile_3);
-		ImageView field4 = findViewById(R.id.tile_4);
-		ImageView field5 = findViewById(R.id.tile_5);
-		ImageView field6 = findViewById(R.id.tile_6);
-		ImageView field7 = findViewById(R.id.tile_7);
-		ImageView field8 = findViewById(R.id.tile_8);
-		ImageView field9 = findViewById(R.id.tile_9);
-		ImageView field10 = findViewById(R.id.tile_10);
-		ImageView field11 = findViewById(R.id.tile_11);
-		ImageView field12 = findViewById(R.id.tile_12);
-		ImageView field13 = findViewById(R.id.tile_13);
-		ImageView field14 = findViewById(R.id.tile_14);
-		ImageView field15 = findViewById(R.id.tile_15);
-		ImageView field16 = findViewById(R.id.tile_16);
-		ImageView field17 = findViewById(R.id.tile_17);
-		ImageView field18 = findViewById(R.id.tile_18);
-		ImageView field19 = findViewById(R.id.tile_19);
-		ImageView field20 = findViewById(R.id.tile_20);
-		ImageView field21 = findViewById(R.id.tile_21);
-		ImageView field22 = findViewById(R.id.tile_22);
-		ImageView field23 = findViewById(R.id.tile_23);
-		ImageView field24 = findViewById(R.id.tile_24);
-		ImageView field25 = findViewById(R.id.tile_25);
-		ImageView field26 = findViewById(R.id.tile_26);
-		ImageView field27 = findViewById(R.id.tile_27);
-		ImageView field28 = findViewById(R.id.tile_28);
-		ImageView field29 = findViewById(R.id.tile_29);
-		ImageView field30 = findViewById(R.id.tile_30);
-		ImageView field31 = findViewById(R.id.tile_31);
-		ImageView field32 = findViewById(R.id.tile_32);
-		ImageView field33 = findViewById(R.id.tile_33);
-		ImageView field34 = findViewById(R.id.tile_34);
-		ImageView field35 = findViewById(R.id.tile_35);
-		ImageView field36 = findViewById(R.id.tile_36);
-		ImageView field37 = findViewById(R.id.tile_37);
-		ImageView field38 = findViewById(R.id.tile_38);
-		ImageView field39 = findViewById(R.id.tile_39);
-
-		return new ImageView[]{field0, field1, field2, field3, field4, field5, field6, field7, field8, field9, field10, field11, field12, field13, field14, field15, field16, field17, field18, field19, field20, field21, field22, field23, field24, field25, field26, field27, field28, field29, field30, field31, field32, field33, field34, field35, field36, field37, field38, field39};
-
-	}
-
-	@Override
-	protected void onCreate(Bundle savedInstanceState){
-		super.onCreate(savedInstanceState);
-		setContentView(R.layout.activity_old2);
-
-		final ImageView[] fields = initializeUI();
-
-		final Player p = new Player("Wutzi", 1000, new GamePiece("shoe"), 0);
-
-
-		chanceCards = new ChanceCardDeck();
-		communityCards = new CommunityCardDeck();
-		chanceCards.initializeDeck();
-		communityCards.initializeDeck();
-		chanceCardProcessor = new ChanceCardProcessor();
-
-
-		gameboard = new Gameboard(getApplicationContext());
-		gameboard.gameboardArray[0] = new GamePiece("Player 1");
-		button_rollDice = findViewById(R.id.button_roll_dice);
-		view_numberDice = findViewById(R.id.view_number_dice);
-		view_position = findViewById(R.id.number_playerposition);
-		view_balance = findViewById(R.id.text_balance);
-		view_balance.setText("Balance: " + p.getBalance());		button_rollDice.setOnClickListener(new View.OnClickListener(){
-			@Override
-			public void onClick(View v){
-				amount = dice.roll();
-
-				gameboard.move("Player 1", amount);
-
-				p.setCurrentPosition(gameboard.getPosition("Player 1"));
-				checkPlayersPosition(p);
-
-
-
-				view_numberDice.setText(Integer.toString(amount));
-				view_position.setText(Integer.toString(gameboard.getPosition("Player 1")));
-
-
-				findViewById(R.id.playericon).setX(fields[gameboard.getPosition("Player 1")].getX());
-				findViewById(R.id.playericon).setY(fields[gameboard.getPosition("Player 1")].getY());
-			}
-		});
-	}
-
-	public void checkPlayersPosition(final Player player) {
-		GameTile currentTile = gameboard.gameTiles.get(player.getCurrentPosition());
-
-		//i know not that performant, but currently i only want to proof our concept.
-		if (currentTile instanceof Street) {
-			final Street street = (Street) currentTile;
-			if (street.getOwner() == null) {
-				AlertDialog dialog = new AlertDialog.Builder(OldActivity2.this).create();
-				dialog.setTitle("Buy Deed?");
-				dialog.setMessage("You are about to buy " + "'" + street.getName() + "' for €" + street.getPrice());
-				dialog.setButton(AlertDialog.BUTTON_NEGATIVE, "NO", new DialogInterface.OnClickListener(){
-					@Override
-					public void onClick(DialogInterface dialog, int which){
-						dialog.dismiss();
-						//TODO: START AUCTION - OR NOT
-					}
-				});
-				dialog.setButton(AlertDialog.BUTTON_POSITIVE, "YES", new DialogInterface.OnClickListener(){
-					@Override
-					public void onClick(DialogInterface dialog, int which){
-						performAcquiringDeed(street, player);
-					}
-				});
-
-				dialog.show();
-			}
-
-
-		}else if (currentTile instanceof Railroad) {
-
-		}else if (currentTile instanceof Utility) {
-
-		}else if (currentTile instanceof CommunityCard){
-			CommunityCard communityCard = communityCards.getNextCard();
-			AlertDialog dialog = new AlertDialog.Builder(OldActivity2.this).create();
-			dialog.setTitle("Community Card!");
-			dialog.setMessage(communityCard.getDescription());
-			dialog.setButton(AlertDialog.BUTTON_NEUTRAL, "OK", new DialogInterface.OnClickListener() {
-				@Override
-				public void onClick(DialogInterface dialog, int which) {
-					dialog.dismiss();
-				}
-			});
-			dialog.show();
-
-		}else if (currentTile instanceof ChanceCard){
-			ChanceCard chanceCard = chanceCards.getNextCard();
-			AlertDialog dialog = new AlertDialog.Builder(OldActivity2.this).create();
-			dialog.setTitle("Chance Card!");
-			dialog.setMessage(chanceCard.getDescription());
-			dialog.setButton(AlertDialog.BUTTON_NEUTRAL, "OK", new DialogInterface.OnClickListener() {
-				@Override
-				public void onClick(DialogInterface dialog, int which) {
-					dialog.dismiss();
-				}
-			});
-			dialog.show();
-
-			chanceCardProcessor.performAction(player, chanceCard);
-			view_balance.setText("Balance: " + player.getBalance());
-		}
-	}
-
-	public void performAcquiringDeed(Deed deed, Player player) {
-		if (deed.getPrice() <= player.getBalance()) {
-			//PLAYER CAN BUY IT
-			int newBalance = player.getBalance() - deed.getPrice();
-			player.updateBalance(-deed.getPrice()); // updated method now subtracts the input from the current balance
-			player.addDeedToPlayer(deed);
-
-			view_balance.setText("Balance: " + player.getBalance());
-		}
-		//TODO: HANDLE CASE IF PLAYER HAS NOT ENOUGH MONEY
-	}
-}
-
